Revert "feat: added rsc support changes next (#1063)"
This reverts commit 93f437ef1d1ff229807b21f44607bf69fc5d7096.
