+++ packages/react-native-css-interop/src/runtime/components.ts
@@ -1,5 +1,3 @@
-"use client";
-
 import {
   ActivityIndicator,
   FlatList,
